restructure module doc strings
