+++ sandbox_request/__main__.py
@@ -1,6 +1,3 @@
-"""
-    Module main.py
-"""
 # Copyright 2021 Universität Tübingen, DKFZ and EMBL
 # for the German Human Genome-Phenome Archive (GHGA)
 #
@@ -15,6 +12,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""Entrypoint of the package"""
 
 from ghga_service_chassis_lib.api import run_server
 from sandbox_request.config import get_config
