+++ sandbox_request/dao/db_connect.py
@@ -1,7 +1,3 @@
-"""
-    Module database.py
-"""
-
 # Copyright 2021 Universität Tübingen, DKFZ and EMBL
 # for the German Human Genome-Phenome Archive (GHGA)
 #
@@ -16,6 +12,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""
+    Database connection handling
+"""
 
 import motor.motor_asyncio
