+++ sandbox_request/routes/__init__.py
@@ -1,6 +1,3 @@
-"""
-    Module __init__.py
-"""
 # Copyright 2021 Universität Tübingen, DKFZ and EMBL
 # for the German Human Genome-Phenome Archive (GHGA)
 #
@@ -15,3 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""
+    This package defines routes/endpoints of the Restful API
+"""
